eval/pyproject.toml
[project]
name = "ditto-eval"
version = "0.0.1"
description = "Evaluation harness for agent memory systems. Runs LongMemEval, BEAM, Ditto-Provenance-Bench, Ditto-Isolation-Bench against pluggable backends."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Unspecified" }
authors = [{ name = "stubbi" }]

dependencies = [
    "click>=8.1",
    "pydantic>=2.7",
    "pyyaml>=6.0",
    "httpx>=0.27",
    "numpy>=1.26",
]

[project.optional-dependencies]
mem0 = ["mem0ai>=0.1"]
zep  = ["zep-python>=2.0"]
dev  = ["pytest>=8.0", "pytest-asyncio>=0.23", "ruff>=0.4", "mypy>=1.10"]

[project.scripts]
ditto-eval = "ditto_eval.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["ditto_eval"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
